chore: add fonts.css to nuxt styles folder
